+++ source/libnemo-private/nemo-global-preferences.h
@@ -216,6 +216,8 @@ enum
 #define NEMO_PREFERENCES_LIST_VIEW_DEFAULT_COLUMN_ORDER		"default-column-order"
 #define NEMO_PREFERENCES_LIST_VIEW_ENABLE_EXPANSION         "enable-folder-expansion"
 #define NEMO_PREFERENCES_LIST_VIEW_COLUMN_FIT_PERCENT       "column-fit-percent"
+#define NEMO_PREFERENCES_LIST_VIEW_ROW_SHADING             "row-shading"
+#define NEMO_PREFERENCES_LIST_VIEW_ROW_SHADING_COLOR       "row-shading-color"
 
 #define NEMO_PREFERENCES_MAX_THUMBNAIL_THREADS "thumbnail-threads"