+++ source/src/nemo-file-management-properties.c
@@ -84,6 +84,7 @@
 #define NEMO_FILE_MANAGEMENT_PROPERTIES_SWAP_TRASH_DELETE "swap_trash_binding_checkbutton"
 #define NEMO_FILE_MANAGEMENT_PROPERTIES_OPEN_NEW_WINDOW_WIDGET "new_window_checkbutton"
 #define NEMO_FILE_MANAGEMENT_PROPERTIES_SHOW_LIST_VIEW_EXPANDERS_WIDGET "list_view_show_expanders_checkbutton"
+#define NEMO_FILE_MANAGEMENT_PROPERTIES_LIST_VIEW_ROW_SHADING_WIDGET "list_view_row_shading_checkbutton"
 
 #define NEMO_FILE_MANAGEMENT_PROPERTIES_SHOW_PREVIOUS_ICON_TOOLBAR_WIDGET "show_previous_icon_toolbar_togglebutton"
 #define NEMO_FILE_MANAGEMENT_PROPERTIES_SHOW_NEXT_ICON_TOOLBAR_WIDGET "show_next_icon_toolbar_togglebutton"
@@ -1585,6 +1586,10 @@ nemo_file_management_properties_dialog_setup (GtkBuilder  *builder,
     bind_builder_bool (builder, nemo_list_view_preferences,
                        NEMO_FILE_MANAGEMENT_PROPERTIES_SHOW_LIST_VIEW_EXPANDERS_WIDGET,
                        NEMO_PREFERENCES_LIST_VIEW_ENABLE_EXPANSION);
+
+    bind_builder_bool (builder, nemo_list_view_preferences,
+                       NEMO_FILE_MANAGEMENT_PROPERTIES_LIST_VIEW_ROW_SHADING_WIDGET,
+                       NEMO_PREFERENCES_LIST_VIEW_ROW_SHADING);
 
     setup_tooltip_items (builder);
     connect_tooltip_items (builder);